InferenceBasedImportance/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
scikit-learn==1.3.2
tqdm==4.66.1
Pillow==10.1.0
tensorboard==2.17.0
torchmetrics==1.1.0
timm==0.9.0
pandas==2.1.3
